login page overhaul with flexbox
