+++ Saucier720-app/src/app/login/FEC/login-form/login-form.component.scss
@@ -1,0 +1,39 @@
+.form-container {
+    display: flex;
+    flex-wrap: wrap;
+    flex-direction: column;
+    align-items: center;
+    padding: 10px;
+    border: 1px solid black;
+  }
+
+  .form-header {
+    margin-bottom: 20px;
+    text-align: center;
+  }
+
+.input {
+  padding-right: 40px;
+
+  &-wrapper {
+    display: flex;
+    align-items: center;
+    position: relative;
+  }
+
+  &-field {
+    position: relative;
+  }
+
+  }
+
+.password-button {
+  position: absolute;
+  right: 0;
+  padding: 0 10px;
+  }
+
+.submit-button {
+  align-self: center;
+  padding: 10px;
+}
