+++ Saucier720-app/src/app/login/login.component.html
@@ -1,42 +1,5 @@
-<div class="outer">
-    <div class="container">
-        <div class="row">
-            <div class="offset-md-2 col-lg-5 col-md-7 offset-lg-4 offset-md-3">
-                <div class="panel border bg-white">
-                    <div class="panel-heading">
-                        <h3 class="pt-3 font-weight-bold">Login</h3>
-                    </div>
-                    <div class="panel-body p-3">
-                        <form>
-                            <div class="form-group pt-3">
-                                <div class="input-field">
-                                    <span class="fa fa-user p-2"></span>
-                                    <input type="text" placeholder="Username" id="username" name="username"[(ngModel)]="username"><br><br>
-                                </div>
-                            </div>
-                            <div class="form-group pt-3">
-                                <div class="input-field">
-                                    <span class="fa fa-lock p-2"></span>
-                                    <input [type]="type" placeholder="Password" id="password" name="password"[(ngModel)]="password"><br><br>
-                                    <span (click)="hideShowPass()" class="fa {{eyeIcon}}"></span>
-                                </div>
-                            </div>
-                            <div class="form-inline fr mt-4">
-                                <a id="forget" class="font-weight-bold">Forget Password</a>
-                            </div>
-                            <button type="submit" (click)="login()" class="btn btn-primary btn-block mt-4 w-100">Submit</button>
-                            <div class="text-center pt-4 text-muted">Not signed up? <a id="forget" routerLink="/Signup">Signup</a></div>
-                        </form>
-                    </div>
-                        <div class="mx-3 my-2 py-2 bordert">
-                            <div class="text-center py-3">
-                                <a href="https://www.google.com" target="_blank" class="px-2"><img src="https://www.freepnglogos.com/uploads/google-logo-png/google-logo-png-google-logos-vector-eps-cdr-svg-download-10.png" alt=""></a>
-                                <a href="https://www.github.com" target="_blank" class="px-2"><img src="https://www.freepnglogos.com/uploads/512x512-logo-png/512x512-logo-github-icon-35.png" alt=""></a>
-                            
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
+<div class="page">
+    <div class="page-body">
+        <app-login-form></app-login-form>
     </div>
 </div>
